requirements.txt
kfp==1.8.22
pandas==2.0.2
google-cloud-bigquery==3.11.1
google-cloud-aiplatform==1.25.0
google-cloud-storage==2.9.0